third_party/blink/web_tests/platform/linux/virtual/scalefactor150/fast/hidpi/static/data-suggestion-picker-appearance-expected.txt

CONSOLE ERROR: Uncaught ReferenceError: test_driver is not defined
PASS successfullyParsed is true

TEST COMPLETE
